Split functions to libraries
Suprisingly, everything still works
