+++ .medkit/index.php
@@ -15,8 +15,8 @@
 	else if (isset($_COOKIE['username']) && isset($_COOKIE['password'])){
 		if (correct_password($_COOKIE['username'], $_COOKIE['password'])) {
 			$_SESSION['username'] = $_COOKIE['username'];
-			$_SESSION["groupID"] = users_get_group_id($_SESSION['username']);
-			$_SESSION["groupName"] = groups_get_name($_SESSION["groupID"]);
+			$_SESSION["groupID"] = users_get_last_group_id($_SESSION['username']);
+			$_SESSION["groupName"] = groups_get_selected_name($_SESSION["groupID"]);
 			header("Location: home.php");
 			exit();
 		}
@@ -37,8 +37,8 @@
 			if(correct_password($username, $password)) {
 				
 				$_SESSION['username'] = $username;
-				$_SESSION["groupID"] = users_get_group_id($_SESSION['username']);
-				$_SESSION["groupName"] = groups_get_name($_SESSION["groupID"]);
+				$_SESSION["groupID"] = users_get_last_group_id($_SESSION['username']);
+				$_SESSION["groupName"] = groups_get_selected_name($_SESSION["groupID"]);
 
 				if (drugs_overdue_check_date($_SESSION["groupID"])){
 					$_SESSION['drugsOverdueModal'] = "show";
